Kill testDisk and the "injection hack"
Summary:
I like the idea of swapping out a "fake" disk, but doing it through Buck introduces a lot of complexity (which you can see on the "red" side of this diff).

In general, I just don't like using the linker for dependency injection like this. It's really quite inflexible, since you need to create a separate binary every time you want to swap out an implementation. We end up with weird APIs where the "test disk" has a "skip mocking" mode that uses the real disk anyway. Ugh!

I'm a fan of fakes and dependency injection, but I'd much rather see us inject at runtime instead of compile time, using functors or just passing arguments to functions or whatever.

Furthermore, I'm not convinced we were not really making good use of fake anyway. We used the "test injector," which pulls in the TestDisk module, for a handful of tests, but those tests don't seem to interact much with the file system anyway.

Reviewed By: mroch

Differential Revision: D39490515

fbshipit-source-id: 39db159e14f153b53fa1643ebf1dcf84bec02471