eliminate stat's during file walk
Summary:
on some filesystems, `readdir` populates `d_type` to tell you if the entry is a file, directory, symlink, etc.

when it exists, we can avoid calling stat() on each entry.

`Sys.readdir` (which calls the C `readdir` repeatedly and returns an array) and `Unix.readdir` (which calls the C `readdir` once) both just return the dirent's name only, so we have to implement our own `readdir` C call to also return the type.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D39438013

fbshipit-source-id: cb7e7d1bad76b9625879a5acef4bdb5ea6fb0c84