Stop emitting missing-local-annot errors for function parameters in statement.ml
Summary:
Env_resolution has to emit missing-local-annot errors for unannotated function parameters, because it might be the last time we visit them due to env_resolution caching.

Now that the reordered-env is the default and the old env dead, we will always run env_resolution, so the same errors we emit in statement.ml becomes redundant. The previous few diffs have fixed final few inconsistencies between the two, so that in this env, we can finally remove these redundant errors!

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39481678

fbshipit-source-id: de3fdeb3b10e90cde58e0f196d7e7becf79db105