Reverse ResolveSpreadT(ResolveSpreadsToMultiflowSubtypeFull)
Summary:
This use_t is relatively common and is created when subtyping functions when one has a spread argument. Since the use_t will never trigger, we can instead use the funtype stored in it to infer a good type. In this diff, I turn the arguments into a tuple and use that as the pinned type.

This use_t has many possible states, but in practice the one where there is only a spread argument is most common, which is why I use `[]` for `rrt_resolved` and `rrt_unresolved`, which indicates that the only parameter to the function was the spread param.

Reviewed By: panagosg7

Differential Revision: D38672170

fbshipit-source-id: 94071630dd8417d15bcd786597a74d3c661b6df9