Fix incremental GC work accounting
Summary:
The incremental GC has an API `collect_slice` which takes an integer "work"
parameter. This controls how much work the GC does before returning.

Before this diff, we consumed our work budget by decrementing by 1 whenever we
marked or swept an object. In effect, the work budget is a number of objects.

This diff uses a different approach: the work budget is the number of words.
That is, if you pass work=100, then the GC will mark or sweep up to 100 words
from the heap.

This new definition is more useful. We can use the size-based work budget to
ensure that the GC keeps pace with allocations.

The new definition is also a bug fix. The implementation of partial-object
marking assumes that the work budget is word-based. Partial-object marking works
by comparing the number of fields with the remaining work budget, and only
scanning `min(num_fields, work)` number of fields.

However, remember that `work` was an *object* counter, not a *word* counter. The
units are wrong. We would mark `work` words, but only decrement `work` by 1, and
then continue marking when we should have stopped.

Reviewed By: mroch

Differential Revision: D39442320

fbshipit-source-id: 632d3961119bd6c1a2316a2b7ccb708930f9e953