No longer propagating hints in statement.ml
Summary:
There is no more need to pass around hints in statement.ml, because we already have a map of hints in the previous diff.

In this diff, I will query that map in places where we still care about hints in statement.ml, and remove all usages of hints otherwise.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39452440

fbshipit-source-id: aa6e10ffaaafd83680d351a73930ed6a498618d2