[contextual] Introduce `Decomp_Instantiated` decomposition op to produce an instantiated callee hint
Summary:
In this diff, I laid out some initial ground work for integrating contextual typing and implicit instantiation. Although implicit instantiation can happen in a lot of places, I choose to focus on function calls initially.

The core of this diff is the new `Decomp_Instantiated` decomposition operation. On a high level, its purpose can be described by the following pseudocode:

```
Decomp_Instantiated callee_t { return_hint_t; arg_ts } {
  if callee_t is polymorphic {
    arg_ts = synthesize(arg_list)
    callee_t = resolve_overload(callee_t, arg_ts)
    subst_map = solve_targs(callee_t, arg_ts, return_hint_t)
    return subst(callee_t, subst_map)
  } else {
    return call_t
  }
}
```

In this diff and the next diff, I will focus mostly on only one use case first: using the return hint to provide better hints on arguments. With this use case in mind and the fact that overload support and synthesis mode is not implemented yet, here are some temporary design choices I made:

1. We will still do a synthesis pass of the argument list, but it will be an extremely dumb one: we just provide an empty tvar for every argument. In the future, it should be replaced with proper synthesis
2. `Decomp_Instantiated` will depend on return hint only. This choice holds for now, since in this diff and the next diff, we are not implementing left-to-right proposal yet. We are temporarily just using the combination of the type of callee and the return hint as the base hint for all arguments in a call. In the future, this restriction will be lifted, and we will record an index in `Decomp_Instantiated`, so we know what are the first `n` number of arguments we need to depend on, for the hint on each argument.
3. return hint and argument list attached to `Decomp_Instantiated` will be a lazy structure. This is a permanent choice. Since `Decomp_Instantiated` is created during syntactic analysis time, where we don't know whether a function is polymorphic, we risk doing unnecessary work of evaluating return hint and synthesizing argument list that we don't need. Thus, we make it a lazy structure, and only force its evaluation when we know the callee is polymorphic.

In this diff, I make the evaluation of `Decomp_Instantiated` return the base hint for now. An initial rough implementation will be done in the next diff.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39214265

fbshipit-source-id: 2aa72747c2a2ff93e6107a2a611372ef63902d0c