Synthesize object literals with annotated methods
Summary:
A somewhat common pattern is object literals that behave as a "bag of functions":
```
const toolkit = {
  doAThing(): number { return toolkit.doSomethingElse() },
  doSomethingElse(): number { ... }
}
```
Currently, any such object literal that refers to itself is considered unsynthesizable, and requires an annotation to be placed on the variable declaration under LTI--but such annotations are pretty gnarly, and if each of the methods were annotated, in theory unnecessary.

This diff implements support for such objects to be considered annotated if each of their methods are themselves fully annotated. It does so by creating a special name def for these objects, and constructing them within env_resolution.ml rather than through statement.ml

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39515088

fbshipit-source-id: 4c83b2bb24ecfe3cad38896b809bde585cc5bbf6