remove unused watchman clock from API
Summary:
the latest watchman clock is an implementation detail that leaks out when fetching the mergebase, but it's unused. this removes it from the API and adds a comment attempting to justify why we don't use it (not sure if that was the original intent but it seems ok).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39629676

fbshipit-source-id: 9e94e77a843b93dacbbd958080923b15c726177f