fix missing recheck when Watchman restarts
Summary:
when watchman restarts, we set `missed_changes = true` to signify that we need to recheck not only the files watchman reports have changed locally, but also all of the previously-modified files (aka focused files) since they may no longer be modified.

we only cancel a recheck if there's something new to include. it's possible that we could be in the middle of rechecking some files when watchman restarts. if the files that have changed locally are a subset of the files we're already rechecking, we currently wouldn't think we need to cancel the recheck, so we wouldn't recheck the previously-modified files.

now we'll cancel the recheck if `missed_changes` changes.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39349029

fbshipit-source-id: ec3e0bcc9e43e4a6787dab7cd1c33669c8532127