Invalidate refinements from LHS of `and` expressions if RHS havocs
Summary:
Currently Flow doesn't handle havocing correctly within `and` expressions--the following snippet does not error at present:

```
let x: number | string = 42;

function havoc() {
    x = 42;
}

(typeof x === 'string') && havoc() && (x: string);
```

The reason for this is that, when inspecting a logical refinement, we capture the refinements from the LHS of an && before we typecheck the RHS. If the RHS havocs, that won't affect the captured refinements from the LHS--and when we merge the LHS and RHS refinements, nothing will prevent the LHS refinements from being live.

This diff addresses this by two changes. First, when we havoc now, in addition to havocing the `env_state.env`, we also go through all the `latest_refinements` and havoc them as well--by which I mean, we filter the applied refinements to be only those names that *weren't* havocked, as well as clearing the changeset and all heap refinements.

With this change, we can now capture the refinements from the LHS refinement scope *after* the RHS has fired in the case where the RHS only runs if the LHS is truthy (the && case). This later-captured set of refinements now accounts for the havoc having occurred in the RHS.

Reviewed By: jbrown215

Differential Revision: D39590622

fbshipit-source-id: 1d5bdb3251b0266947d92b0cab62fea3f0912d41