[implicit-instantiation] Make `in_implicit_instantiation` flag only care about whether we are in implicit instantiation
Summary:
D39482936 (https://github.com/facebook/flow/commit/7a1aa65031f0f95521fd6cf6d7f908909f562e36) introduced the flag to detect whether we are in post inference implicit instantiation. The goal is to change the behavior of flow_js such that:

1. Avoid adding upper bounds from tparam bounds
2. Stop flowing TypeDestructorTriggerT lower bounds in implicit instantiations.

In the diff, we agreed that 1 is helpful for all implicit instantiation purposes, so that it is helpful for this flag to target all implicit instantiation, not only during the post inference phase.

If we want to enable some behavior specifically for the post inference pass, we can detect that using Context.PostInference.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39584304

fbshipit-source-id: a4f53c5586e9ad68c30c5d8e0f5bff9dc6371d87