concretize the callee of a CallT constraint
Summary:
In upcoming work it will be useful to have a facility that will help us expose
a concrete constructor (such as a `FunT`, an `IntersectionT` or a `PolyT`) that
is about to be called. This is because we are planning to have different behaviors
depending on the kind of callee type.

Today getting this constructor is not straightforward even with fully-resolved tvars.
When calling
* `EvalT`s they first need to be evaluated,
* `ClassT`s are replaced with the statics of the class,
* `ObjT`s are replaced with the callable property if existing,
* `GenericT`s are replaced with their bounds,

etc.

This diff adds a variant to the `CallT` constraint, encoded in the  `call_action`
field as `ConcretizeCallee`. The other variant `Funcalltype` represents the
existing behavior. `ConcretizeCallee` performs all the operations listed above,
which aim at bringing one of the primitive call-type forms to the head of the callee
type. Instead of performing the call (ie matching actual to formal parameters)
it returns the form of the lower bound of that constraint to a designated tvar that
expects this result.

This is the equivalent of the `NoMethodAction` variant of `method_action`s
whose purpose is to return the type of the receiving object called property in
method calls.

Reviewed By: SamChou19815

Differential Revision: D39639306

fbshipit-source-id: 3455ff228282ceca6e4bc09682f9f7c0eb1c7157