Factor out tvar_resolver to its own module
Summary:
We want to be able to seal tvars at the boundary of `Statement.expression` calls. While we are not there yet, we can first factor out tvar resolver into its own module to prepare for the future step. This would allow type_hint to use it without going through a callback. (If we directly do `Env_resolution.TvarResolver.resolved_t` in type hints, it will create a cycles in modules.)

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39640553

fbshipit-source-id: 8dc1fecad7dc0ee5c23599a3046b4fe89f957b79