Enforce constructor must return void
Summary:
In this diff, I dropped the support for non-void return in class constructors. I achieved it by unifying return_t with VoidT with a return use_op, so that you can have nice error messages. This helps us to easily compute the signature of class constructors: we don't need return annotation even if the constructor throws, since it will always be pinned to void.

Changelog: [errors]
Previously we support a weird JS runtime rule for constructor returns when you call `new SomeClass`:

- If the returned value is primitive, then it's ignored and the instance will be returned instead.
- If the returned value is object like, then the return object will be the eval result of `new SomeClass`.

This is confusing, and we dropped support for this feature in this release. Now we enforce that all class constructors must return void.

Reviewed By: samwgoldman

Differential Revision: D39278909

fbshipit-source-id: 92e256a02a3d3162a1e58c2210ad3760dd8882dd