[contextual] Turn off empty tvar errors when we are under synthesis mode
Summary:
During intermediate steps of evaluating hints, we shouldn't care about whether a tvar is resolved or not, because an unresolved tvar might not appear in the final result at the end of the chain. Instead, we should only care about this at the end of hint evaluation.

In this diff, I restructured the hint evaluation process as follows:
1. When evaluating a hint decomposition, we will be in synthesis mode with empty tvar errors turned off. Before doing the decomposition, we fully resolve the input.
2. In the final step before we output the result, we will fully resolve the output, and return None if any tvars in the result are unresolved.

Changelog: [internal]

(Note: this ignores all push blocking failures!)

Reviewed By: panagosg7

Differential Revision: D39640552

fbshipit-source-id: 0750cff307d946cc163436a048acf09c2467889e