Constructors for tvars as placeholders
Summary:
Changelog: [internal]

In this diff, I added constructors for placeholders. When creating a tvar as a placeholder with `Tvar.mk_placeholder`. The tvar id will be automatically added the placeholder set in the context. The placeholder set in context will be useful in the future. e.g. when we are fully resolving types in synthesis mode, we should ignore those placeholder tvars.

Reviewed By: SamChou19815

Differential Revision: D38076505

fbshipit-source-id: bfa31e7e2e5a0b95d9112312846f1510a6a327b2