Fix for havocing shadowed variables by keeping a stack for each variable in the env
Summary:
Previously, a shadowed variable would be immune to havocing, even if at runtime it would be mutated by a function call: e.g.,

```
declare var invariant: any

let aa: string | number = 42;
let havoc = function () { aa = 'a' };
invariant(typeof aa === 'number');

{
  let aa: string | number = 42;
  invariant(typeof aa === 'number');
  havoc();
  (aa: number); // fine
}

(aa: number); // fail
```
This is because a shadowed variable doesn't exist as an accessible thing when we havoc--it's stored in the call stack by the saved result of `push_env`, and is then restored when we `pop_env`.

This diff makes shadowed variables still exist in the env state by modifying the env to contain a stack of values for each variable. The top of the stack is what's actually in scope, and the rest of the stack is shadowed. When we havoc from a call (but importantly, not from entering a function scope) we havoc everything in the stack, inlcuding things that are currently shadowed. We don't havoc everything when we enter a function scope, because this havocing shouldn't affect values outside of the function scope.

Reviewed By: SamChou19815

Differential Revision: D39678284

fbshipit-source-id: 10b86ef00a1172ce6cba9fed350dc62bfd10ef18