[codemod] Convert `React$Element<typeof SomeComponent>` to `React.MixedElement` under `--generalize-react-mixed-element`
Summary:
The logic is first introduced in D13906974 (took me a while to find, so I copied the rationale in that diff into the comment). Converting such type to React.Node loses context that something is a react element rather than the more general node, which causes the logic I introduced in D39756499 (https://github.com/facebook/flow/commit/fb2ccba74c796d39d60dbe57325c55e594d21034) to be useless.

Now that we have React.MixedElement, converting such type into React.MixedElement is also a valid choice.

I make the change conditional on `--generalize-react-mixed-element`, so that we can experiment with it with codemod and see whether it makes sense to make the behavior the new default.

Reviewed By: pieterv

Differential Revision: D39782427

fbshipit-source-id: ba4448920416fcce7cc7a277d9b37d50d8df31b1