hints for intersection type calls
Summary:
This diff implements the decomposition of `Decomp_Instantiated` for `IntersectionT` inputs.

First off, to detect that this is indeed an overloaded call, it uses the `ConcretizeCallee`
call action that was introduced in D39639306 (https://github.com/facebook/flow/commit/ea87d5da0d8d892be13e1c7edbf3fe992bd71929), and checks that the toplevel constructor
is an `IntersectionT`.

Once that's established, we *synthesize* types for the arguments, that is infer types without
passing any hints to them, and using "placeholder" tvars instead of parameters and return
types for unannotated lambdas. We use these argument types to fire a `CallT` constraint,
 whose goal is to determine the overload that should be selected overall.

The speculation mechanism today does not have a way to return the winning overload. We
add this capability by introducing a mutable field in the `CallT-funcalltype` structure. This
records the state (`speculation_hint_state`) of speculation in a call to an intersection type in
synthesis mode. The important bits to record here are the type of the successful overload,
and the path of speculation ids that led us there. On subsequent successes, we'll make sure
that the speculation id belongs to the successful path. This means that we're in the process
of popping branches ("firing actions") in the case of nested speculation. If not we irreversibly
end up in an "invalid" state.

An example of getting multiple valid successful branches are nested intersections
```
type B = (tag: 'bool', cb: (x: boolean) => void) => void;
type S = (tag: 'str', cb: (x: string) => void) => void;
type N = (tag: 'num', cb: (x: number) => void) => void;
type O = (tag: 'obj', cb: (x: {}) => void) => void;

declare var foo: B & (S & (N & O));
foo('num', (x) => { (x: string); }); // TODO error
```
The successful branches are the ones that correspond to types `N`, `(N & O)`, and
`(S & (N & O))`. We only want to record the most helpful type `N`.

An example of invalid subsequent successful branches is
```
declare function foo(x: number, f: (number) => void): void;
declare function foo(x: string, f: (string) => void): void;
declare var x: number | string;
foo(x, y => {});
```
Flow will break up the union first, before processing the intersection call. The two union
members in this example lead to different overload selection results, belonging to sibling
branches, so one does not subsume the other.

It is also possible that we don't add an entry for a specific CallT (`Speculation_hint_unset`).
This corresponds to the case where no overload was selected, either because they all failed,
or we couldn't select one due to ambiguity.

 ---

Note that this process is only used for creating hints for call arguments. Once that's done, we
continue with performing the call as usual. As an optimization, we can in the future cache the
result of overload resolution to be reused later.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39738929

fbshipit-source-id: ce2838a1a07294be99241bf1c7c1f1814d5d0577