Do not clear files during ensure parsed
Summary:
In lazy mode, we use "ensure parsed" in a couple places:
1. For direct dependent files, before we try to re-resolve requires, which needs access to parse artifacts.
2. Later, for all dependent files (for merge + check) and their dependencies (for merge)

The `ensure_parsed` API calls into the same parsing reducer as `reparse` does, but it is a bit of a special case:
1. We only call the reducer for files which have not yet been parsed
2. We know the set of files is expected to exist and we have a content hash
3. We only care about the subset of files which is different from expected, or not found

We use this information to cancel a recheck, but we do not expect that it will make any changes to the heap that need to be undone.

Before this change, for the "file not found" case, we always called `Parsing_heaps.clear_file` for missing files. This has a number of side effects:

1. It modifies the file's parse entity to None, indicating a missing file
2. It modifies the file's haste ent to None, which affects later commit_modules phases
3. It modifies reverse dependencies, removing the file from it's dependencies' dependents sets

None of these side effects should happen for ensure parsed. When we cancel a recheck, they will not necessarily be rolled back, because the files are not necessarily in the recheck mutator's changed file set.

This is a patch, but a better, forthcoming fix is to implement ensure_parsed as a special case without reusing the same parsing reducer as the rest of rechecks. Even the cases where we actually want to parse and update the heap (when the file exists and has a matching digest), it can be done more efficiently.

Reviewed By: mroch

Differential Revision:
D39877267 (https://github.com/facebook/flow/commit/a75d5b792ee71684594e8baf929121718e8637ac)

------------------------------------------------------------------------
(from 7fb7f3e97fc0198acd87b54830e7737d30ff9cc2)

fbshipit-source-id: e5613ade6c93eb5c5e0c5746e0cc33a4a59f130a