Boilerplate for implicit instantiation codemod
Summary:
This diff adds all the relevant boilerplate code for the implicit instantiation codemod. I used the same flags as the `useState` codemod, since that codemod is very similar in flavor to this one.

I initialize the error set and have a pass where calls consult the error set, but do not act on them yet. That will happen in D39867503

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39867505

fbshipit-source-id: 82616e3c9a18307b819d1405853f88a36ad620ef