[codemods] Add flag to save implicit instantiation results
Summary:
The implicit instantiation codemod will need to insert inferred types for implicit instantiations. Since the type arguments are not in the ast, we need to keep around a map on the side to record the types.

That map will likely take up a large amount of memory. To avoid saving those results in non-codemod contexts, I'm adding a flag to options to save the implicit instantiation results. Note that there is no way to enable this flag in user land-- only in code as we do in `src/commands/codemodCommand.ml`

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39867502

fbshipit-source-id: be779b0bd9e3670419299b402c1982c4ad9c0103