Add inferred targ results to context to query in codemods
Summary:
Since the targs are not part of the type ast, we need to keep them in a map on the side so that we can query them for the implicit instantiation codemod. This diff does exactly that and consults the `save_implicit_instantiation_results` flag to avoid doing so outside of the codemod.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39867504

fbshipit-source-id: 71cd7df26f5e9f99bc31ddcd58416335dcd972d1