Contextual typing for empty arrays
Summary:
This diff builds contextual typing support for arrays by reading the hint from the environment and flowing the array's type into it.

This diff also ensures that empty arrays' tvars are inhabited in all cases, including having no hint and having hints that don't provide a type: for example, `var x: { ... } = { a: [] }` provides a hint for `[]`, but that hint can't be resolved to a type.

Reviewed By: SamChou19815

Differential Revision: D39677093

fbshipit-source-id: fc2f52f5b393bc119230af4c230704018bc0c170