[contextual] Apply the same left-to-right and overload resolution algorithm to constructor call
Summary:
In previous diffs, we implemented left-to-right algo and overload resolution for function and method calls. Constructor calls are currently left behind, which is fixed in this diff.

The changes on the name-def side is very straightforward, we call `visit_call_arguments` on the constructor call args, so that the arguments will go through the same `Decomp_Instantiated` that does left-to-right and overload resolution. However, with this change alone, things still don't work, and I found an existing bug of `Decomp_CallNew` and `Decomp_CallSuper`.

In both `Decomp_CallNew` and `Decomp_CallSuper`, we just try to get the method named "constructor" on the instance. We doesn't check whether the instance is polymorphic at all! This results in some type like `(GenericT<T>) => void` instead of `PolyT(((GenericT<T>) => void, [T])`. Therefore, I made an additional check on whether the input type is a PolyT, if so, we will do the decomposition on t_out of PolyT, and then wrap the result with the PolyT wrapper again.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39835180

fbshipit-source-id: cb9c4ddaf0d3a23ef6230e93b21a8441dce3db60