[lti] First cut of left-to-right algorithm
Summary:
This diff implements the left-to-right algorithm as we discussed.

For implicit_instantiation_hints, we will keep track of the index of the argument we are checking. In dependency analysis, we declare that we only depend on the expressions before the index, and in hint evaluation, we only replace the types of expressions with their checked types for expressions before the index.

In the end, we will be able to type check the pipe example, which requires the left-to-right algorithm to work!

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39680657

fbshipit-source-id: fe841d269cd85c8f686b14d58387c7739ea4e99c