Reverse ObjKitT(ReactConfig)
Summary:
This was the second most common use_t in upper bounds that I found. It is emitted when flowing from a class or a function to an AbstractComponent in a polymorphic signature. The upper bound on the tout is the props type of the component, which is a very reasonable type to infer in the vast majority of cases, so I use that to pin the type. Note that this isn't as precise as possible for function components with defaults, as illustrated in the added test.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38672420

fbshipit-source-id: 5d33b63873355e681d4e6a77e4306410b07466a1