add --no- variants of truthy flags
Summary:
adds a `no_arg` flag type that for flag `--foo` also defines a flag `--no-foo`. neither take a value like `--foo false`.

I used this to add a `--no-abstract-locations` flag, because that setting defaults to true so as is, there was no way to disable it without touching the flowconfig.

I implemented it this way, instead of just adding another `flag "--no-abstract-locations" ...`, so that you can pass both `--no-foo` and `--foo` and have the last one win. if they were separate flags the ordering gets lost. that's useful for scripting.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40031515

fbshipit-source-id: f2c6bcfec2ce45c5db51f583e4323b91f17b8607