Rewrite hh_add to avoid double-word CAS, remove libatomic dependency
Summary:
Before this change, we used a DWCAS operation to replace the entire hash slot at
once. It turns out this was unnecessary and we can use two single-word CAS
operations instead.

The key insight is that we can serialize access to the hash field and addr field
separately, in that order. Two threads racing to take a slot will try to take
the hash field first.

When a thead races to take a slot, there are three possible outcomes:
1. It wins the race and writes it's hash value to the slot
2. It fails the race, but the winning thread writes the same hash value
3. If fails the race, and the winning thread writes some other value

In cases (1) and (2), the thread will continue and try to take the addr field.
In case (3), we continue probing.

An interesting interleaving is now possible: when two threads race to write to
the same slot with the same hash, the thread that wins the race on the hash
field might differ from the thread that wins the race on the addr field.

Reviewed By: mroch

Differential Revision: D40076918

fbshipit-source-id: 1f8e5977695a5976b3a9f3ee7a16271e9c5f3a67