Do not trigger contextual typing when `array_literal_providers` flag is on but LTI flag is off
Summary:
When `array_literal_providers` is on and the array is has hint, we will run contextual typing, which is only designed to be run together with LTI. In this diff, I will do nothing when the array literal as hint but LTI flag is not on. This ensures that `array_literal_providers` flag can be independently turned on.

Changelog: [internal]

Reviewed By: pieterv

Differential Revision: D40122234

fbshipit-source-id: cf3dac6b9b9d56914dcadd6f0041dd672b2a0f39