Speed up reindexing
Summary:
The server environment maintains an exports index which needs to be updated
during recheck.

The index is a very large O(repo) size map, the "update" is a very small
O(changes) map. When updating the large map, we need to be careful to do
O(update) amount of work.

Before this change, we used the `merge` operation to combine the index with the
update, but merge does work proportional to the larger map. This diff re-writes
the update functions to be biased toward the update map.

Instead of `merge` over both maps, we `fold` over the O(changes) update map.

Reviewed By: mroch

Differential Revision: D40092166

fbshipit-source-id: bc06b3fba70e780afe1f3e9a36fafe39e6435831