Fix incremental crash when removing old dependents
Summary:
The first three steps of recheck are: parse, commit modules, resolve requires

Parse finds "dirty modules" which are modules whose provider may have changed.
In commit modules, we go through dirty modules and update their providers, as
needed.

Before this diff, when a module had no providers and no dependents, we would
remove that module from the hash table (at the end of the transaction).

However, this decision to remove the module is premature. In the resolve
requires step, which happens after commit modules, we might add a new dependent.
In this case, we would end up removing a module that has dependents, which leads
to crashes.

While working on incremental reverse dependencies, I realized a similar issue in
remove_old_dependent. We don't have enough information to decide, because a new
dependent might be added later.

For the fix, I decided to simply not remove modules with no providers and no
dependents. As noted in the x-ref'ed TODO comments, we could maintain a set of
candidate modules to remove, then check at the end of the transaction before
removing them.

It is not really unreasonable to leak these modules, though. A haste module with
no providers or dependents only takes up 40 bytes of space + the space to store
the haste name. A file module only requires 32 bytes of space.

Reviewed By: mroch

Differential Revision: D40080085

fbshipit-source-id: d9c2ca2202e1c8dab05ecd7b13cfbaaa15a8839f