Resolve tvars at expression boundaries
Summary:
This diff implements sealing of tvars at the boundary of calls to Statement.expression, except for expressions that have a hint availalble to them (in which case, the hint will provide the external type and the hint is already guaranteed to be resolved).

This diff does not require that the expression being resolved has lower bounds--we don't emit errors on unbounded tvars, but rather we just resolve them to empty. This has a minor effect on the error diff, because if an expression is resolved to empty after a call to `Statement`, and then is resolved again by env resolution, the second attempt at resolving it won't see it as an empty tvar that it would signal an error for.

On our test cases, the same number of test cases fail both with and without this change, but the specific errors do change somewhat. For the most part, this is due to repositioned errors and loss of use ops. There are a few more interesting cases though, most of which seem to be improvements. For example, the following case:

```
enum M {
  A,
}
enum N {
  A,
}
function f8(x: M | N, X: typeof M | typeof N) {
  switch (x) { // Error
    case X.A: break; // Error
  }
}
```
Here there are currently the following errors:
```

Error --------------------------------------------------------------------------------------- exhaustive-check.js:188:11

Incomplete exhaustive check: the member `A` of enum `M` [1] has not been considered in check of `x`.
[invalid-exhaustive-check]

   exhaustive-check.js:188:11
   188|   switch (x) { // Error
                  ^

References:
   exhaustive-check.js:187:16
   187| function f8(x: M | N, X: typeof M | typeof N) {
                       ^ [1]

Error --------------------------------------------------------------------------------------- exhaustive-check.js:188:11

Incomplete exhaustive check: the member `A` of enum `N` [1] has not been considered in check of `x`.
[invalid-exhaustive-check]

   exhaustive-check.js:188:11
   188|   switch (x) { // Error
                  ^

References:
   exhaustive-check.js:187:20
   187| function f8(x: M | N, X: typeof M | typeof N) {
                           ^ [1]

Error --------------------------------------------------------------------------------------- exhaustive-check.js:189:10

Invalid check of case test against switch discriminant [1] because `M` [2] is incompatible with `N` [3].
[incompatible-type]

   exhaustive-check.js:189:10
   189|     case X.A: break; // Error
                 ^^^

References:
   exhaustive-check.js:188:11
   188|   switch (x) { // Error
                  ^ [1]
   exhaustive-check.js:187:16
   187| function f8(x: M | N, X: typeof M | typeof N) {
                       ^ [2]
   exhaustive-check.js:187:37
   187| function f8(x: M | N, X: typeof M | typeof N) {
                                            ^^^^^^^^ [3]

Error --------------------------------------------------------------------------------------- exhaustive-check.js:189:10

Invalid check of case test against switch discriminant [1] because `N` [2] is incompatible with `M` [3].
[incompatible-type]

   exhaustive-check.js:189:10
   189|     case X.A: break; // Error
                 ^^^

References:
   exhaustive-check.js:188:11
   188|   switch (x) { // Error
                  ^ [1]
   exhaustive-check.js:187:20
   187| function f8(x: M | N, X: typeof M | typeof N) {
                           ^ [2]
   exhaustive-check.js:187:26
   187| function f8(x: M | N, X: typeof M | typeof N) {
                                 ^^^^^^^^ [3]

```

With this change, the second two errors (pointing at `X.A`) go away. This is likely because these errors occurred due to a tvar gaining `M` and `N` as both upper and lower bounds, and errors resulted from this. With this change, tvars don't have the opportunity to gain bounds in the same way, and we only get the first two errors.

Reviewed By: SamChou19815

Differential Revision: D39677094

fbshipit-source-id: 525ae8784fe5d2641e0936918322eb0b320c63f7